Prototyp 3, Version 8.1
//		+ Stopp der Simulation
//			Unsere Simulation soll ab einem gewissen Punkt (bei Anzahl X an
Infizierten) gestoppt werden
//			dann soll sie einen kurzen Moment lang wirken und übergehen in den
Erkenntnisscreen
//
//		+ Anpassung der Gesamtform der Stadt und des Kindergartens
//			Es soll kein Hexagon mehr sein, da das zu viel Aufmerksamkeit auf
sich zieht
//			und möglicherweise die Nutzer verwirrt. Stattdessen soll es eine
Kreisannäherung
//			werden.
//
//		+ TimeDisplay (Zeitleiste)

+++ prototyp3/virus.pde
@@ -122,13 +122,13 @@ class Virus{
 	}
 
 
-	public void moveTo( int _x, int _y, float _time ){
+	public void moveTo( float _x, float _y, float _time ){
 
 		moveList.add( new VirusTarget( new PVector(_x,_y), _time, 0 ) );
 
 	}
 
-	public void moveTo( int _x, int _y, float _time, int _animationType ){
+	public void moveTo( float _x, float _y, float _time, int _animationType ){
 
 		moveList.add( new VirusTarget( new PVector(_x,_y), _time, _animationType ) );
